+++ src/main/java/lazyae2/network/ModNetwork.java
@@ -28,5 +28,6 @@ public final class ModNetwork {
         CHANNEL.registerMessage(PacketTerminalAmount.Handler.class, PacketTerminalAmount.class, 4, Side.SERVER);
         CHANNEL.registerMessage(PacketTerminalFilter.Handler.class, PacketTerminalFilter.class, 5, Side.SERVER);
         CHANNEL.registerMessage(PacketAssemblerPatterns.Handler.class, PacketAssemblerPatterns.class, 6, Side.CLIENT);
+        CHANNEL.registerMessage(PacketAssemblerWork.Handler.class, PacketAssemblerWork.class, 7, Side.CLIENT);
     }
 }